Remove old wordSection components. Begin transitioning word-view-edit to new lexicalSections
